requirements.txt
fastapi>=0.115.0
httpx>=0.28.0
pandas>=2.2.0
python-dotenv>=1.0.0
requests>=2.32.0
sqlalchemy>=2.0.0
streamlit>=1.42.0
uvicorn>=0.34.0
